[FIX] 테스트 오류나던 스케줄러 부분 fix
